feat: 블랙잭 제약 사항 검증 기능 구현

+++ src/main/java/blackjack/dto/ProfitDto.java
@@ -2,5 +2,5 @@ package blackjack.dto;
 
 import java.util.Map;
 
-public record ProfitDto(Map<String, Integer> profit) {
+public record ProfitDto(Map<String, Double> profit) {
 }
